Refactor Asset Inventory to make room for grouping
