+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/components/asset_inventory_loading.tsx
@@ -6,7 +6,7 @@
  */
 import React from 'react';
 import { EuiFlexGroup, EuiFlexItem, EuiLoadingLogo, EuiSpacer } from '@elastic/eui';
-import { InventoryTitle } from './inventory_title';
+import { AssetInventoryTitle } from './asset_inventory_title';
 import { CenteredWrapper } from './onboarding/centered_wrapper';
 import { TEST_SUBJ_LOADING } from '../constants';
 
@@ -16,7 +16,7 @@ import { TEST_SUBJ_LOADING } from '../constants';
 export const AssetInventoryLoading = () => (
   <EuiFlexGroup>
     <EuiFlexItem>
-      <InventoryTitle />
+      <AssetInventoryTitle />
       <EuiSpacer size="l" />
       <CenteredWrapper>
         <EuiLoadingLogo logo="logoSecurity" size="xl" data-test-subj={TEST_SUBJ_LOADING} />
