+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/components/asset_inventory_search_bar.tsx
@@ -12,17 +12,15 @@ import type { Filter } from '@kbn/es-query';
 import { useKibana } from '../../common/lib/kibana';
 import { FiltersGlobal } from '../../common/components/filters_global/filters_global';
 import { useDataViewContext } from '../hooks/data_view_context';
-import type { AssetsURLQuery } from '../hooks/use_asset_inventory_data_table';
+import type { AssetsURLQuery } from '../hooks/use_asset_inventory_url_state/use_asset_inventory_url_state';
 
 interface AssetInventorySearchBarProps {
   setQuery(v: Partial<AssetsURLQuery>): void;
-  loading: boolean;
   placeholder?: string;
   query: AssetsURLQuery;
 }
 
 export const AssetInventorySearchBar = ({
-  loading,
   query,
   setQuery,
   placeholder = i18n.translate(
@@ -32,14 +30,13 @@ export const AssetInventorySearchBar = ({
     }
   ),
 }: AssetInventorySearchBarProps) => {
+  const { dataView } = useDataViewContext();
   const { euiTheme } = useEuiTheme();
   const {
     unifiedSearch: {
       ui: { SearchBar },
     },
   } = useKibana().services;
-
-  const { dataView } = useDataViewContext();
 
   return (
     <FiltersGlobal>
@@ -49,7 +46,6 @@ export const AssetInventorySearchBar = ({
           showFilterBar={false}
           showQueryInput={true}
           showDatePicker={false}
-          isLoading={loading}
           indexPatterns={[dataView]}
           onQuerySubmit={setQuery}
           onFiltersUpdated={(filters: Filter[]) => setQuery({ filters })}
