+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/components/asset_inventory_title.tsx
@@ -7,15 +7,18 @@
 import React from 'react';
 import { EuiTitle } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
+import { TechnicalPreviewBadge } from './technical_preview_badge';
+import { TEST_SUBJ_PAGE_TITLE } from '../constants';
 
-export const InventoryTitle = () => {
+export const AssetInventoryTitle = () => {
   return (
-    <EuiTitle size="l" data-test-subj="inventory-title">
+    <EuiTitle size="l" data-test-subj={TEST_SUBJ_PAGE_TITLE}>
       <h1>
         <FormattedMessage
-          id="xpack.securitySolution.assetInventory.inventoryTitle"
+          id="xpack.securitySolution.assetInventory.title"
           defaultMessage="Inventory"
         />
+        <TechnicalPreviewBadge />
       </h1>
     </EuiTitle>
   );
