+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/components/filters/asset_inventory_rule_type_ids.ts
