+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/hooks/use_data_view.ts
