+++ x-pack/solutions/security/plugins/security_solution/public/asset_inventory/routes.tsx
@@ -14,7 +14,7 @@ import { SecuritySolutionPageWrapper } from '../common/components/page_wrapper';
 import { PluginTemplateWrapper } from '../common/components/plugin_template_wrapper';
 import { SecurityRoutePageWrapper } from '../common/components/security_route_page_wrapper';
 import { DataViewContext } from './hooks/data_view_context';
-import { useDataView } from './hooks/use_asset_inventory_data_table/use_data_view';
+import { useDataView } from './hooks/use_data_view';
 import { AssetInventoryLoading } from './components/asset_inventory_loading';
 import { ASSET_INVENTORY_INDEX_PATTERN } from './constants';
